zhidun-project/backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
python-multipart==0.0.6
openai>=1.0.0
python-dotenv
edge-tts>=6.1.0
